+++ MarketPlace924/Service/IDummyWalletService.cs
@@ -12,7 +12,7 @@ namespace MarketPlace924.Service
         /// </summary>
         /// <param name="userId">The ID of the user whose wallet balance to retrieve.</param>
         /// <returns>A task representing the asynchronous operation that returns the wallet balance.</returns>
-        Task<float> GetWalletBalanceAsync(int userId);
+        Task<double> GetWalletBalanceAsync(int userId);
 
         /// <summary>
         /// Updates the wallet balance for the specified user.
@@ -20,6 +20,6 @@ namespace MarketPlace924.Service
         /// <param name="userId">The ID of the user whose wallet balance to update.</param>
         /// <param name="newBalance">The new wallet balance.</param>
         /// <returns>A task representing the asynchronous operation.</returns>
-        Task UpdateWalletBalance(int userId, float newBalance);
+        Task UpdateWalletBalance(int userId, double newBalance);
     }
 }
