+++ MarketPlace924/Service/IOrderSummaryService.cs
@@ -24,7 +24,7 @@ namespace MarketPlace924.Service
         /// <param name="additionalInfo">Additional information about the order.</param>
         /// <param name="contractDetails">Contract details for borrowed items.</param>
         /// <returns>A task representing the asynchronous operation.</returns>
-        Task UpdateOrderSummaryAsync(int id, float subtotal, float warrantyTax, float deliveryFee, float finalTotal,
+        Task UpdateOrderSummaryAsync(int id, double subtotal, double warrantyTax, double deliveryFee, double finalTotal,
                                      string fullName, string email, string phoneNumber, string address,
                                      string postalCode, string additionalInfo, string contractDetails);
