+++ Server/Repository/DummyWalletRepository.cs
@@ -28,7 +28,7 @@ namespace Server.Repository
         }
 
         /// <inheritdoc/>
-        public async Task<float> GetWalletBalanceAsync(int userId)
+        public async Task<double> GetWalletBalanceAsync(int userId)
         {
             DummyWalletEntity wallet = await this.dbContext.DummyWallets.FirstOrDefaultAsync(dw => dw.BuyerId == userId)
                                     ?? throw new Exception($"GetWalletBalanceAsync: Wallet not found for buyer with ID: {userId}");
@@ -36,7 +36,7 @@ namespace Server.Repository
         }
 
         /// <inheritdoc/>
-        public async Task UpdateWalletBalance(int userId, float newBalance)
+        public async Task UpdateWalletBalance(int userId, double newBalance)
         {
             DummyWalletEntity wallet = await this.dbContext.DummyWallets.FirstOrDefaultAsync(dw => dw.BuyerId == userId)
                                     ?? throw new Exception($"UpdateWalletBalance: Wallet not found for buyer with ID: {userId}");
